Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -3,7 +3,7 @@
 
 
 int main() {
-    float a;
+    double a;
     scanf("%lf", &a);
     printf("You entered: %.4f", a);
     return 0;
